+++ standingssync/__init__.py
@@ -1,4 +1,6 @@
+"""Alliance Auth app for cloning alliance standings and war targets to alts."""
+
 default_app_config = "standingssync.apps.SyncAltContactsConfig"
 
-__version__ = "1.8.1"
+__version__ = "1.9.0"
 __title__ = "Standings Sync"