+++ standingssync/managers.py
@@ -1,6 +1,8 @@
+"""Managers for standingssync."""
+
 import datetime as dt
 from collections import defaultdict
-from typing import Dict, Optional, Set, Tuple
+from typing import Any, Dict, Set, Tuple
 
 from django.contrib.auth.models import User
 from django.db import models, transaction
@@ -19,7 +21,7 @@ logger = LoggerAddTag(get_extension_logger(__name__), __title__)
 
 
 class EveContactQuerySet(models.QuerySet):
-    def grouped_by_standing(self) -> Dict[int, models.Model]:
+    def grouped_by_standing(self) -> Dict[int, Any]:
         """Group alliance contacts by standing and convert into sorted dict."""
         contacts_by_standing = defaultdict(set)
         for contact in self.all():
@@ -66,7 +68,7 @@ class EveWarQuerySet(models.QuerySet):
         """Add is_active field to queryset. Requires prior annotation of state."""
         return self.annotate(
             is_active=Case(
-                When(state__in=self.model.State.active_states, then=Value(True)),
+                When(state__in=self.model.State.active_states(), then=Value(True)),
                 default=Value(False),
             )
         )
@@ -124,7 +126,7 @@ class EveWarManagerBase(models.Manager):
 
         return EveEntity.objects.filter(id__in=war_target_ids)
 
-    def update_or_create_from_esi(self, id: int) -> Tuple[models.Model, bool]:
+    def update_or_create_from_esi(self, id: int) -> Tuple[Any, bool]:
         """Updates existing or creates new objects from ESI with given ID."""
 
         logger.info("Retrieving war details for ID %s", id)
@@ -152,7 +154,7 @@ class EveWarManagerBase(models.Manager):
             )
             war.allies.clear()
             if war_info.get("allies"):
-                for ally_info in war_info.get("allies"):
+                for ally_info in war_info.get("allies", []):
                     try:
                         ally = self._get_or_create_eve_entity_from_participant(
                             ally_info
@@ -187,7 +189,7 @@ EveWarManager = EveWarManagerBase.from_queryset(EveWarQuerySet)
 
 
 class SyncManagerManager(models.Manager):
-    def fetch_for_user(self, user: User) -> Optional[models.Model]:
+    def fetch_for_user(self, user: User) -> Any:
         """Fetch sync manager for given user. Return None if no match is found."""
         if not user.profile.main_character:
             return None