+++ standingssync/tests/core/test_esi_contacts.py
@@ -8,7 +8,15 @@ from standingssync.core.esi_contacts import (
     EsiContactsContainer,
 )
 
-from ..factories import EsiContactFactory, EsiContactLabelFactory, EveContactFactory
+from ..factories import (
+    EsiContactFactory,
+    EsiContactLabelFactory,
+    EveContactFactory,
+    EveEntityAllianceFactory,
+    EveEntityCharacterFactory,
+    EveEntityCorporationFactory,
+    EveEntityFactionFactory,
+)
 
 MODULE_PATH = "standingssync.core.esi_contacts"
 WAR_TARGET_LABEL = "WAR TARGETS"
@@ -25,7 +33,7 @@ class TestEsiContact(NoSocketsTestCase):
 
     def test_should_create_new_2(self):
         # when
-        obj = EsiContact(1001, "character", 5.0)
+        obj = EsiContact(1001, "character", 5.0)  # type: ignore
         # then
         self.assertEqual(obj.contact_id, 1001)
         self.assertEqual(obj.contact_type, EsiContact.ContactType.CHARACTER)
@@ -34,11 +42,28 @@ class TestEsiContact(NoSocketsTestCase):
     def test_should_create_new_3(self):
         # when/then
         with self.assertRaises(ValueError):
-            EsiContact(1001, "xyz", 5.0)
+            EsiContact(1001, "xyz", 5.0)  # type: ignore
+
+    def test_should_create_all_types(self):
+        # given
+        params = [
+            ("character", EsiContact.ContactType.CHARACTER),
+            ("corporation", EsiContact.ContactType.CORPORATION),
+            ("alliance", EsiContact.ContactType.ALLIANCE),
+            ("faction", EsiContact.ContactType.FACTION),
+        ]
+        for input, expected in params:
+            with self.subTest(input=input):
+                # when
+                obj = EsiContact(1001, input, 5.0)
+                # then
+                self.assertEqual(obj.contact_id, 1001)
+                self.assertEqual(obj.contact_type, expected)
+                self.assertEqual(obj.standing, 5.0)
 
     def test_should_clone_contact_1(self):
         # given
-        a = EsiContact(1, "character", 5.0, [1, 2])
+        a = EsiContact(1, EsiContact.ContactType.CHARACTER, 5.0, frozenset([1, 2]))
         # when
         b = a.clone()
         # then
@@ -46,7 +71,7 @@ class TestEsiContact(NoSocketsTestCase):
 
     def test_should_clone_contact_2(self):
         # given
-        a = EsiContact(1, "character", 5.0, [1, 2])
+        a = EsiContact(1, EsiContact.ContactType.CHARACTER, 5.0, frozenset([1, 2]))
         # when
         b = a.clone(standing=-10)
         # then
@@ -55,31 +80,31 @@ class TestEsiContact(NoSocketsTestCase):
         self.assertEqual(b.label_ids, a.label_ids)
         self.assertEqual(b.standing, -10)
 
-    def test_should_create_from_dict_1(self):
+    def test_should_create_from_esi_character(self):
         # given
         esi_dict = {"contact_id": 1, "contact_type": "character", "standing": 5.0}
         # when
         obj = EsiContact.from_esi_dict(esi_dict)
         # then
-        self.assertEqual(obj, EsiContact(1, "character", 5.0))
+        self.assertEqual(obj, EsiContact(1, EsiContact.ContactType.CHARACTER, 5.0))
 
-    def test_should_create_from_dict_2(self):
+    def test_should_create_from_esi_corporation(self):
         # given
         esi_dict = {"contact_id": 1, "contact_type": "corporation", "standing": 5.0}
         # when
         obj = EsiContact.from_esi_dict(esi_dict)
         # then
-        self.assertEqual(obj, EsiContact(1, "corporation", 5.0))
+        self.assertEqual(obj, EsiContact(1, EsiContact.ContactType.CORPORATION, 5.0))
 
-    def test_should_create_from_dict_3(self):
+    def test_should_create_from_esi_alliance(self):
         # given
         esi_dict = {"contact_id": 1, "contact_type": "alliance", "standing": 5.0}
         # when
         obj = EsiContact.from_esi_dict(esi_dict)
         # then
-        self.assertEqual(obj, EsiContact(1, "alliance", 5.0))
+        self.assertEqual(obj, EsiContact(1, EsiContact.ContactType.ALLIANCE, 5.0))
 
-    def test_should_create_from_dict_4(self):
+    def test_should_create_from_esi_contact_when_labels_are_none(self):
         # given
         esi_dict = {
             "contact_id": 1,
@@ -90,7 +115,30 @@ class TestEsiContact(NoSocketsTestCase):
         # when
         obj = EsiContact.from_esi_dict(esi_dict)
         # then
-        self.assertEqual(obj, EsiContact(1, "alliance", 5.0))
+        self.assertEqual(obj, EsiContact(1, EsiContact.ContactType.ALLIANCE, 5.0))
+
+    def test_should_create_from_esi_faction(self):
+        # given
+        esi_dict = {"contact_id": 1, "contact_type": "faction", "standing": 5.0}
+        # when
+        obj = EsiContact.from_esi_dict(esi_dict)
+        # then
+        self.assertEqual(obj, EsiContact(1, EsiContact.ContactType.FACTION, 5.0))
+
+    def test_should_create_from_eve_entities(self):
+        # given
+        params = [
+            (EveEntityCharacterFactory(), EsiContact.ContactType.CHARACTER),
+            (EveEntityCorporationFactory(), EsiContact.ContactType.CORPORATION),
+            (EveEntityAllianceFactory(), EsiContact.ContactType.ALLIANCE),
+            (EveEntityFactionFactory(), EsiContact.ContactType.FACTION),
+        ]
+        for eve_entity, expected in params:
+            with self.subTest(category=eve_entity.category):
+                # when
+                obj = EsiContact.from_eve_entity(eve_entity=eve_entity, standing=5.0)
+                # then
+                self.assertEqual(obj, EsiContact(eve_entity.id, expected, 5.0))
 
 
 class TestEsiContactLabel(NoSocketsTestCase):