+++ standingssync/tests/test_tasks.py
@@ -10,8 +10,9 @@ from app_utils.testing import (
     generate_invalid_pk,
 )
 
-from .. import tasks
-from ..models import SyncedCharacter, SyncManager
+from standingssync import tasks
+from standingssync.models import SyncedCharacter, SyncManager
+
 from .factories import EveContactFactory, SyncedCharacterFactory, SyncManagerFactory
 from .utils import ALLIANCE_CONTACTS, LoadTestDataMixin
 
@@ -105,7 +106,7 @@ class TestCharacterSync(LoadTestDataMixin, NoSocketsTestCase):
         # given
         mock_update.return_value = True
         # when
-        tasks.run_character_sync(self.synced_character_2)
+        tasks.run_character_sync(self.synced_character_2.pk)
         # then
         self.assertTrue(mock_update.called)
 
@@ -115,7 +116,7 @@ class TestCharacterSync(LoadTestDataMixin, NoSocketsTestCase):
         mock_update.side_effect = RuntimeError
         # when
         with self.assertRaises(RuntimeError):
-            tasks.run_character_sync(self.synced_character_2)
+            tasks.run_character_sync(self.synced_character_2.pk)
 
 
 @patch(TASKS_PATH + ".run_character_sync")